Move normalization unit tests to integration tests (too slow)
