+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/edge_cases_catalog.json
