+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/edge_cases_catalog_expected_nested.json
