+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/edge_cases_catalog_expected_top_level.json
