+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/edge_cases_catalog_expected_top_level_postgres.json
@@ -1,0 +1,14 @@
+{
+  "_airbyte_schema_test": {
+    "postgres_has_a_64_ch__inations_are_fine_ab1": "postgres_has_a_64_ch__inations_are_fine_ab1_f6d",
+    "postgres_has_a_64_ch__inations_are_fine_ab1_611": "postgres_has_a_64_ch__inations_are_fine_ab1_611_967",
+    "postgres_has_a_64_ch__inations_are_fine_ab2": "postgres_has_a_64_ch__inations_are_fine_ab2_f6d",
+    "postgres_has_a_64_ch__inations_are_fine_ab2_611": "postgres_has_a_64_ch__inations_are_fine_ab2_611_967",
+    "postgres_has_a_64_ch__inations_are_fine_ab3": "postgres_has_a_64_ch__inations_are_fine_ab3_f6d",
+    "postgres_has_a_64_ch__inations_are_fine_ab3_611": "postgres_has_a_64_ch__inations_are_fine_ab3_611_967"
+  },
+  "schema_test": {
+    "postgres_has_a_64_ch__destinations_are_fine": "postgres_has_a_64_ch__destinations_are_fine_8ce",
+    "postgres_has_a_64_ch__destinations_are_fine_611": "postgres_has_a_64_ch__destinations_are_fine_611_acc"
+  }
+}
