+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/facebook_catalog.json
