+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/facebook_catalog_expected_nested_bigquery.json
