+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/facebook_catalog_expected_top_level.json
