+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/stripe_catalog.json
