+++ airbyte-integrations/bases/base-normalization/integration_tests/resources/naming/stripe_catalog_expected_nested_bigquery.json
